requirements.txt
mysql-connector-python
scrapy
dateparser
python-dotenv
